version 1  uploaded
This is a python script that will do the following:
Allow the user to choose  one of these options
 1) IP address lookup tools
 2) Threat attribution tools 
 3) Website scanners 
 4) IP Fraud Search and Blacklist Checkers 
 5) Threat and Vulnerability Research
 6) User agent parsers - Sources used to parse HTTP user agent string 
 7) MAC Address lookup tools
 8) Commonly used IP sites used to make Reports 

If the user picks option 1, they are prompted to enter a vaild public IP address and once it is valid the browser will actomatically open most links with inputed IP address results. These websites will each be opened in a separate tab.

ip2location.com/demo
https://search.censys.io/hosts
https://www.shodan.io/host/
 

If the user picks option 2, they are prompted to enter a vaild public IP address and once it is valid the browser will actomatically open most links with inputed IP address results. These websites will each be opened in a separate tab.
https://otx.alienvault.com
https://www.virustotal.com/gui/home/search
https://metadefender.opswat.com/ 
https://www.abuseipdb.com/ 
https://exchange.xforce.ibmcloud.com/
https://community.riskiq.com/research
https://pulsedive.com

If the user picks option 3, they'll be prompted to enter a vaild URL and once it is valid actomatically open their default web browser and paste the inputed URL in the searchbar of this website 
https://urlscan.io/

If the user picks option 3, they are prompted to enter a vaild public IP address and once it is valid the browser will actomatically open most links with inputed IP address results. These websites will each be opened in a separate tab.

https://mxtoolbox.com/blacklists.aspx 
https://scamalytics.com/ip/ 
https://whatismyipaddress.com/blacklist-check 
https://threatfox.abuse.ch/browse/ 

If the user picks option 5, they are prompted to enter a vaild public IP address and once it is valid the browser will actomatically open most links with inputed IP address results. These websites will each be opened in a separate tab.
https://onyphe.io 
https://vulners.com/ 
https://packetstormsecurity.com/ 


If the user picks option 6 actomatically open their default web browser and open this website  
https://developers.whatismybrowser.com/useragents/parse/#parse-useragent 

If the user picks option 7, they'll be prompted to enter a vaild MAC address and once it is valid actomatically open their default web browser and paste the inputed MAC address in the searchbar of one of these websites while just opening one 
https://dnschecker.org/mac-lookup.php?query=00%3AE0%3A4C%3A00%3A08%3AAD 6698533
https://macvendors.com/